Allow to get Properties from content's predecessors
